+++ server/src/middleware/auth.ts
@@ -1,0 +1,31 @@
+import { Request, Response, NextFunction } from "express";
+
+import jwt from "jsonwebtoken";
+
+interface JwtPayload {
+  username: string;
+}
+
+export const authenticateToken = (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+
+    const secretKey = process.env.JWT_SECRET_KEY || "";
+
+    jwt.verify(token, secretKey, (err, user) => {
+      if (err) {
+        return res.sendStatus(403); // forbidden status
+      }
+      req.user = user as JwtPayload;
+      return next();
+    });
+  } else {
+    res.sendStatus(401); //unauthorized
+  }
+};
