+++ server/src/routes/auth-routes.ts
@@ -1,5 +1,5 @@
 import { Router, Request, Response } from "express";
-import { User } from "../../models/user.js";
+import { User } from "../models/user.js";
 import jwt from "jsonwebtoken";
 import bcrypt from "bcrypt";
